Moved View annotation to javax.mvc package
