+++ api/src/main/java/javax/mvc/View.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package javax.mvc.annotation;
+package javax.mvc;
 
 import javax.mvc.Controller;
 import java.lang.annotation.Documented;
